Fixed issue with the renaming of the api path to utils
