+++ src/app/project/[id]/page.tsx
@@ -1,5 +1,5 @@
 import {DataType} from "@/lib/DataTypes.d";
-import {getLocalData} from "@/api/dataLayer";
+import {getLocalData} from "@/utils/dataLayer";
 import * as motion from "motion/react-client"
 import ProjectModal from "@/app/project/[id]/modal";
